dfst/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
gast==0.6.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.3
keras==2.14.0
kornia==0.8.0
kornia_rs==0.1.8
libclang==18.1.1
loguru==0.7.3
Markdown==3.8
MarkupSafe==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pillow==11.0.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sympy==1.13.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorrt==8.5.3.1
termcolor==3.0.1
tf-keras==2.15.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchgeometry==0.1.2
torchvision==0.21.0+cu118
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0
